Add step/start properties to properly place player in correct tiles  …
After figuring out how big the column and rows are through the engine.js file, that information (col 101 and row 83) was transferred to the Hero constructor to accurately place player in correct tile as well as to move around within the tile properly.

4 new properties were created. 2 for the horizontal and vertical movement of the player (horizonStep/vertStep) and the other 2 as the starting position (startX and startY).

The startX and startY is multiplied by 2 and 5 respectively to place the player in the correct starting tile. It moves the player 2 tiles over and 5 tiles down based on the number declared (101 & 83).

The switch statement was also updated to include the new horizontal and vertical properties. Each time an input is a match, it will subtract or add horizonStep/vertStep accordingly to move the player around the tiles.