Set boundaries/loop where enemy bug can/can't move
The position/tile where enemy bug can move is 101 column which is set to this.step = 101. this.step * 5 is doesn't include the tile the enemy bug starts at. The 5 refers to the 6th tile just outside of the boundaries to make it appear as though the enemy bug left the board.

The if statement includes "this.x += 200 * dt". 200 refers to the speed and dt keeps the speed constant throughout all devices/browsers.

The else statement resets the enemy bug to just outside the board with "this.x = this.step * -1". After this reset, the enemy bug will continue to move through the board as a result of the if statement referenced above.

This loop continues.